Refactor: TaskCard with custom hook
